+++ docs/CHANGELOG.md
@@ -1,4 +1,6 @@
 # Change log
+- 3.26.20 Support for 'modifier' functions inside Refresh actions, like `Actions.refresh({ key: 'drawer', open: value=>!value })`
+- 3.26.0 Support for React Native 0.26
 - 3.22.20 fix more ESLint errors, fix passing leftButtonStyle property for back button
 - 3.22.18 fix some ESLint errors and ignore pop for root scene
 - 3.22.17 allow jump & push together - now you could call Actions.tab2_2() (`tab2_2` is next scene after `tab2`) even if `tab2` is not active